bug fixes, cleaned up some code

+++ routes/resourceRouter.js
@@ -8,7 +8,6 @@ const {StudyResources} = require('../models');
 
 
 router.get('/', (req, res) => {
-    console.log(req.query)
     const filters = {};
     const queryableFields = ['course', 'typeOfResource', 'username', 'title', '_id'];
 
@@ -21,23 +20,19 @@ router.get('/', (req, res) => {
     StudyResources
         .find(filters)
         .exec()
-        .then(Resources => {
-          console.log(Resources)
-          res.json(Resources.map(resource => resource.apiRpr()))})
+        .then(Resources => res.json(Resources.map(resource => resource.apiRpr())))
         .catch(err => {
               console.error(err);
               res.status(500).json({message: 'internal server error'})
         })
-})
+});
 
 
 router.post('/', (req, res) => {
     const todaysDate = new Date().toLocaleString();
     const requiredFields = ['content', 'course', 'title', 'typeOfResource']
 
-    const missingFields = requiredFields.filter(field => {
-        return !field in req.body
-    })
+    const missingFields = requiredFields.filter(field =>  !field in req.body)
 
     if (missingFields.length > 0) {
         const message = `Request is missing ${missingFields.length} fields.`
@@ -90,14 +85,19 @@ router.put('/:id', (req, res) => {
 
 
 router.delete('/:id', (req, res) => {
+    if (!req.params.id) {
+        const message = `Request path is missing request ID.`
+        console.error(message)
+
+        return res.status(400).send(message)
+    }
+
     let resourceId = req.params.id
 
     return StudyResources
     .findOneAndRemove({_id: resourceId})
     .exec()
-    .then((resource) => {
-      console.log(resource)
-      res.status(201).json(resource)})
+    .then((resource) =>  res.status(201).json(resource))
     .catch(err => res.status(500).json({message: 'Internal Server Error'}))
 })
